[WIP] Re-add Husky, consolidate types
